feature: user can register and login

+++ index.js
@@ -1,6 +1,12 @@
 #!/usr/bin/env node
-const { getCommandsDirectory, directoryExists } = require("./lib/dir");
-const { askCommands } = require("./lib/cli");
+const {
+  getCommandsDirectory,
+  getCredentialsDirectory,
+  directoryExists
+} = require("./lib/dir");
+const { askCommands, askCredentials } = require("./lib/cli");
+const { login, listcommand, newcommand, register } = require("./lib/api");
+
 const { helper } = require("./lib/help");
 const fs = require("fs");
 const columnify = require("columnify");
@@ -21,9 +27,22 @@ const getcommands = () => {
   }
 };
 
-if (directoryExists(getCommandsDirectory())) {
+if (
+  directoryExists(getCommandsDirectory()) &&
+  directoryExists(getCredentialsDirectory())
+) {
   helper();
   process.argv.slice(2).forEach(function(cmd) {
+    if (cmd === "register" || cmd === "--register") {
+      askCredentials().then(credentials => {
+        register(credentials);
+      });
+    }
+    if (cmd === "login" || cmd === "--login") {
+      askCredentials().then(credentials => {
+        login(credentials);
+      });
+    }
     if (cmd === "new" || cmd === "--n") {
       askCommands().then(answers => {
         const cmds = getcommands();
@@ -42,16 +61,22 @@ if (directoryExists(getCommandsDirectory())) {
     }
   });
 } else {
-  let cmd = [];
-  fs.mkdirSync(os.homedir() + "/.commands");
-  hide.setAttributesSync(os.homedir() + "/.commands", { IS_HIDDEN: true });
-  fs.writeFile(
-    os.homedir() + "/.commands/cmd.json",
-    JSON.stringify(cmd),
-    { flag: "wx" },
-    err => {
-      if (err) throw err;
-      console.log("bika initialized!");
-    }
-  );
+  if (directoryExists(getCommandsDirectory()) === false) {
+    let cmd = [];
+    fs.mkdirSync(os.homedir() + "/.commands");
+    hide.setAttributesSync(os.homedir() + "/.commands", { IS_HIDDEN: true });
+    fs.writeFile(
+      os.homedir() + "/.commands/cmd.json",
+      JSON.stringify(cmd),
+      { flag: "wx" },
+      err => {
+        if (err) throw err;
+        console.log("bika initialized!");
+      }
+    );
+  }
+  if (directoryExists(getCredentialsDirectory()) === false) {
+    fs.mkdirSync(os.homedir() + "/.bika/");
+    hide.setAttributesSync(os.homedir() + "/.bika", { IS_HIDDEN: true });
+  }
 }
